requirements.txt
# PyTorch for Apple Silicon M4
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0

# Machine Learning and Optimization
optuna>=3.4.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0

# Data handling and HTTP requests
requests>=2.31.0
urllib3>=2.0.0
tqdm>=4.65.0

# Bioinformatics
biopython>=1.81

# Development tools
black>=23.0.0
pytest>=7.4.0
jupyter>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
